test/all: Enable parallelization of tests
This makes tests run far faster.

Various groups of tests for SynapseBridge have been split into separate
files and describe() blocks to aid in parallel testing.
